Move w32subset.h into utils.
