+++ mono/eglib/gmodule-win32.c
@@ -36,7 +36,7 @@
 #include <windows.h>
 #include <psapi.h>
 #include <gmodule.h>
-#include "../w32subset.h"
+#include "../utils/w32subset.h"
 
 #define LIBSUFFIX ".dll"
 #define LIBPREFIX ""
