+++ mono/eglib/gunicode-win32.c
@@ -9,7 +9,7 @@
 
 #define CODESET 1
 #include <windows.h>
-#include "../w32subset.h"
+#include "../utils/w32subset.h"
 
 #if HAVE_API_SUPPORT_WIN32_GET_ACP || HAVE_API_SUPPORT_WIN32_GET_CP_INFO_EX
 extern const char *eg_my_charset;
