+++ mono/metadata/w32socket-internals.h
@@ -23,7 +23,7 @@
 #define socklen_t int
 #endif
 
-#include <mono/w32subset.h>
+#include <mono/utils/w32subset.h>
 
 #if defined(HOST_WIN32) && (HAVE_API_SUPPORT_WIN32_TRANSMIT_FILE || HAVE_API_SUPPORT_WIN32_DISCONNECT_EX)
 #include <mswsock.h>
